+++ CHANGELOG.md
@@ -17,6 +17,8 @@
   properly ([rhbz#2227233])
 - Improved error messages and documentation of `pcs resource move` command
   ([rhbz#2233763])
+- Do not display warning in `pcs status` for expired constraints that were
+  created by moving resources ([rhbz#2111583])
 
 [rhbz#2217850]: https://bugzilla.redhat.com/show_bug.cgi?id=2217850
 [rhbz#2219388]: https://bugzilla.redhat.com/show_bug.cgi?id=2219388
@@ -24,6 +26,7 @@
 [rhbz#2227233]: https://bugzilla.redhat.com/show_bug.cgi?id=2227233
 [rhbz#2227234]: https://bugzilla.redhat.com/show_bug.cgi?id=2227234
 [rhbz#2233763]: https://bugzilla.redhat.com/show_bug.cgi?id=2233763
+[rhbz#2111583]: https://bugzilla.redhat.com/show_bug.cgi?id=2111583
 
 
 ## [0.10.17] - 2023-06-19